pyproject.toml
[tool.poetry]
name = "abby"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"  # 確保你的 Python 版本不變
openai = "^1.57.4"  # 與 langchain 0.3.9 兼容的版本
langchain = "0.3.9"  # 目標版本
langchain-google-genai = "2.0.5"  # 需要與 langchain 0.3.9 兼容的版本
langchain-community = "0.3.8"  # 可能需要同步降低到兼容版本
google-generativeai = "^0.8.0"  # 與 langchain-google-genai 兼容的版本
python-dotenv = "^1.0.1"  # 環境變量讀取工具
pytest = "^8.3.4"  # 測試框架，版本不受影響
flask = "^3.1.0"  # 與 langchain 無關，可保持原版本
pandas = "^2.2.3"  # 與 langchain 無直接關聯，可保持原版本
line-bot-sdk = "^3.14.2"
langchain-openai = "^0.2.14"
pymysql = "^1.1.1"
redis = "^5.2.1"
cryptography = "^44.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
